Implement average importance.
The average importance of a todo item the average of the importance of
its parents (including its own importance).

This requires some data to ship together with each todo item, since the
sorter function needs access to the parents. Because its unaware of
TodoList which maintains the hierarchy, copy a bit of that information
into the Todo itself.