Recalculate text ID when the text changes.
This fixes a bug that occurred with the combination of text IDs and
appending projects of the parent todo item.

When adding the project to a child todo, the text ID was not updated,
and the standard output would show an outdated text ID. The next
invocation of topydo would show the correct ID.

Now the ID is consistent between adding and listing.