Printer methods must always return a string.
There is no need for clients of printers to assume anything about the
format. In this case, the View class assumed that the printer would
return a list of strings, and then the view joined those entries. This
latter step is actually the responsibility of the printer, so move it
there.