Eliminate possibility for multiple todos having the same id: value.
When there is a todo in the list with an id: tag but no child todos,
that value could be reused when creating a new dependency between two
(unrelated) items.

This is because the value of the dangling parent todo is not present in
the dependency graph, but the function generating a new ID depended on
that.

Instead, simply iterate over all available todo items to check whether
the proposed value already exists somewhere.