Make sure that text identifier creation is deterministic.
Text identifiers were calculated with the hash() builtin. The problem is
that hash() returns different values depending on Python's major
version or 32 vs 64 bit. With Python >=3.3, the seed for the hash
function is always different (for security).

Instead of relying on hash(), use the sha1() hash to obtain a text
identifier.