Create test script to run tests.
The trigger is that pylint doesn't support Python 3.2 anymore, and in
that case Travis CI should not run pylint.

This script takes an optional argument for the executable name of the
corresponding Python version (typically just 'python2' and 'python3').
When not given, 'python' is just run, whatever is in your $PATH. The
latter mode is used in Travis CI.