Allow configuration options to be overridden by commandline flags.
When obtaining a configuration object, it is now possible to pass
certain values to be overridden no matter what settings were made in the
configuration files.

This is used to let the -t and -d flags (location of todo.txt and
archive respectively) override anything that is mentioned in
configuration files.

This fixes a bug where the edit subcommand did not respect the -t flag,
because it was simply reading the location from the configuration file.
Now the configuration is overridden when -t is passed.

Thanks to Jacek for reporting the broken 'edit' command.