Prompt mode: reload the todo.txt file if it was modified.
The todo.txt file can be modified while the prompt is running. Make sure
to reload the file when the mtime has changed since the previous load.

This addresses the issue that changes after the 'edit' subcommand were
not visible when returning to the topydo prompt (issue #33).

Also, other changes done while the prompt is running will be picked up.

A safety measure was taken to refuse executing a command when the
todo.txt file was modified while the prompt was visible (i.e. the user
was busy typing a command). By doing so, subcommands such as 'do' and
'del' will only operate on the actual state of the todo.txt file, not
some previous state still in memory.