Introduce `ls -f <format>` to specify the output format.
Instead of creating a subcommand for every possible output format we're
going to think of in the future, just let ListCommand handle the output
format by connecting to the right format printer.

In this commit, the 'ical' subcommand is deprecated, to be removed after
the next release.

The 'json' subcommand has been removed altogether, it existed for only a
few hours.