Some fixes regarding color sequences in todo items.
This is a revert of commit fbed527, which resulted in too many items
after a project/context being highlighted.

Instead, don't print a NEUTRAL_COLOR code at the end of the todo, but
only when we finish applying all the color codes. This makes sure that
the regexp to match a project/context also works at the end of a line.

Also removed some logic to remove multiple sequences of NEUTRAL_COLOR at
the end of the line, this is not necessary.