Refactored the View class and Unicode support cleanups.
This started as a refactoring of the View class: to get rid of the
pretty_print() function and anything releated to printers. So the View
class now solely focusses on sorting and viewing a bunch of todos. It's
up to the users of the View class to retrieve the todos in the view and
pass them on to a Printer instance.

The View class was quite involved with the proper Unicode support that
was added lately. There were still encoded string passed through this
class rather than Unicode strings. Getting rid of this affected
basically all other places in the codebase.

The result is that there are less string conversions; printed todo items
stay Unicode as long as possible, until they're actually written out to
standard output or to a file. I got rid of one text_type() call from the
six library, all @python_2_unicode_compatible decorators are gone and
the tests no longer need the utf8() function because everything is
Unicode now.

Remove utf8 function.