Fix for relative dates not always being calculated properly.
Fell into the trap of calling a function as a default parameter. Because
of this, relative dates used with the 'add' or 'do' command (recurrence)
were based on an offset when topydo was started, not when the subcommand
was executed. Usually these two events occur in the same second and this
bug was unnoticed for a long time. But now with prompt mode, this bug
may surface if you leave running topydo for longer than a day.