Make 'topydo ls foo:somestring' work.
If there was a todo with 'foo:somestring', it wouldn't be returned with
the command above, because somestring couldn't be converted to a date or
an integer, and therefore the OrdinalTagFilter just returns False.

As a final fallback, use a GrepFilter match to do the string comparison
such that the todo item(s) with this key:value combination will be
returned.