Add "revert" command
`topydo revert` will search for backup corresponding with current state
of todo file, and use it to recover previous state. Output will show to
user results of which command had been reverted. If no suitable backup
is found (for example if user edited todo file recently with other
application), no action is taken and user is notified on the output
about lack of corresponding backup in backup file.