Sanitize defaults in topydo.lib.Config._Config
Using defaults as a parameter passed to the constructor of
configparser.ConfigParser results in a presence of **all** options
inside every section. It's very convenient in many use-cases, but not
very much with our configuration model.

After this change only relevant options are present in each section, so:

`topydo.lib.Config.config().has_option('topydo', 'priority_colors')`

will now return `False` as 'priority_colors' resides in 'colorscheme'
section and not in 'topydo'. Before this change it would return `True`.