Remove restriction of equal lines before and after editing
There was a restriction in place that a file should contain the same
number of todo items before and after an edit. This is now removed.

Reason for change is that I lost a lot of edits to a long list of items
and removing one single item from the list. The editor exits and as a
result topydo refused to process it, losing all the changes.

Another reason is that we have `revert` now. In case you severely mess
up, you can always revert.