Don't write to the filesystem when trimming.
ChangeSet._trim may call ChangeSet.delete() several times, which incurs
a disk write every time. This creates a lot of overhead, especially when
lots of revisions are removed.

Instead, leave the responsibility to the caller of _trim to make sure
that the new content is written to disk. In this case, ChangeSet.save()
the only caller, already writes to disk.