Add -i flag to ls to select only those todo items with a specific ID
-i accepts a comma separated list of todo IDs and `ls` will only print
those IDs that were given.

Use case:

I hide most tags from my list for a clearer overview. In order to see
the tags corresponding to an item, I have to open the todo.txt file and
locate the item there.

Together with aliasing and the upcoming feature with formatting, using
the -i flag allows me to show only those todo items with a specific ID
with the full content.

    [ls]
    hide_tags = bunch,of,tags

    [aliases]
    lsdetail = ls -F "%s %K" -i

Then:

    topydo> ls
    |  1| Hidden details

    topydo> lsdetail 1
    Hidden details bunch:1 of:2 tags:3