Add printer for Graphviz Dot format
This generates a graph of todo items, where the edges denote
dependencies between items. The tooltip shows the description (visible
in an SVG by using a decent browser).

Based on the idea that I implemented in topydo's predecessor:
Dot output by todo.txt-tools [1]. There, I implemented it as a
subsubcommand of `dep` to visualize dependencies. This implementation,
however, is part of the `ls` command, where you can use a filter
expression to limit the size of the graph.

The todo.txt-tools implementation also prints a box with a lookup table
of the numbers to descriptions. This is lacking at the moment.

To visualize the todo items of a project, run:

    topydo ls -x -f dot +Project | dot -Tsvg -o Project1.svg

(-x is suggested because parents with unfinished todo items are not
printed by default).

[1] https://github.com/bram85/todo.txt-tools/blob/2f9b2baa6382346c5cd9d2272c61ee0e342e25b6/actions/dep#L162