Move format parsing logic to ListFormatParser
ListFormatParser is now the core of PrettyPrinterFormatFilter, but can
also be easily used in another areas. Whole functionality of
PrettyPrinterAlignFilter was also moved into it.

Main advantage of having separate parser for list_format is possibility
of initializing it and resolving placeholders only once. Previous
implementation did that with each Todo object.