Use creation date as start date when no explicit start date is given
An explicit start date is given with the t: tag (unless configured
otherwise). Together with the due date, the start date is used to
calculate the length of a todo item (in days).

When no explicit start date is given, consider the creation date as the
start date (if present).