Add option to set the due date to sub todo items
Add this to the configuration file:

    [dep]
    set_due_date_to_subtodo: 1

And due dates are automatically added to child items, if the child item
has no due date yet.

Also add a warning when you add a dependency where the due date of the
parent item is before the due date of the child item.