Fix transitive reduction for parent items with multiple children
Given these dependencies:

A id:1
B p:1
C p:1 id:2
D p:1 p:2

A -> B
A -> C
A -> D
C -> D

The relation A -> D is superfluous (A -> C, C -> D). However, the p:1
tag in D would not be removed, because relation 1 still exists (A -> B).

Therefore a new function is added that identifies p: tags which have no
corresponding edge in the internal dependency graph (rather than looking
at the value of the edge).