Add filters for completion date and creation date
Addresses issue #86 which requested filters on completion date. The
following tags are recognized:

* create
* created
* creation
* complete
* completed
* completion

They are specific ordinal tag filters, except that these tags don't
exist in a todo item.

Filtering on completion date does not work when completed items are
automatically archived. Archived items are not read when invoking `ls`,
so the done.txt file should be read as the main file instead:

    topydo -t done.txt ls complete:today

Also in this commit, decouple ExpressionCommand from Filter class a bit.

The expression class shouldn't bother which expressions belong to which
filter classes, let the Filter module handle that.