Don't write backups, todo files or update columns for read only commands
This addresses some slugginess that I encounter in the column UI. For
example, when doing an `ls` on the commandline, formerly it would
execute:

* Perform backups;
* Write the todo file (+ archive if applicable);
* Refresh all the columns;

All in all this takes quite a while.

CLIApplicationBase::_post_execute will now reset the dirty flag of the
todo list. Then writing stuff and refreshing columns will only occur
when the todo list is dirty.

Updating columns can still be made smarter. When an action is executed
on a specific todo item, only update those columns that contain that
item, or are related to it through dependencies.