Store tags as a dictionary
Store tags in a dictionary where the key is the tag name and the value
is a list of values for that tag.

This makes tag operations much quicker (especially TodoBase::has_tag()
and TodoBase::tag_values()). They made quite a dent in execution time
with heavy todo.txt files with many tags, due to the numerous list
comprehensions that took place.