Introduce -C flag to steer color output
The -C flag takes 0, 1, 16, 256, auto, yes, true, no, false as possible
values.

If auto, then colors are enabled unless the output is not connected to a
TTY and is likely not to understand ANSI escape codes. 'auto' is the
default value.

The config().colors() function either returns 0, 16 or 256 for the
possible colors that can be displayed.

These values are also recognized by the colors option in the
configuration file.

This fixes issue #111.