Embed colors in strings
TopydoString is a string descendent that maintains information where
colors should be applied in a string. A user interface can pass a
function that transforms a color to a color representation that is
relevant there. That way, the topydo library does not make any
assumptions anymore on where colors should be applied.