Show colors in the console
This is achieved using abstract colors for projects, contexts, etc.. The
Color filter for the pretty printer has no knowledge of colors, it just
marks a certain text as 'project color'. All this information is stored
in the TopydoString. The CLI will convert the colored parts in the
string with the corresponding ANSI codes, the column UI will convert the
string to an appropriate markup for the Text widget.

Also, palette items are now specified using an enum-like class
(PaletteItem) to reduce hidden dependencies of the palette item names.

Moreover, ConsoleWidget is now a urwid.Pile with Text widgets, instead
of one single Text widget with all text. This also allows for doing
operations on lines in the output in the future (mark a line and hit
complete, for instance).