Fix crash when completing/deleting items after an edit.
The crash occurs when using text-based identifiers and performing an edit on
selected items (instead of the whole todo.txt file).

If you edit the text of a child item (having a p tag), then the item is deleted
using the TodoListBase.delete() instead of TodoList.delete(). But
this messes up the dependency administration, still thinking that the removed
todo item is still in the list. When you call the children of any todo item
(part of the deletion/completion step), it will return a stale Todo item, whose
number() can not be found anymore in the TodoListBase. That makes it crash.