Fix crash when completing a child item after a 'dep add'
When doing a 'dep add' command, the parent lookup dictionary was not kept up to
date. Therefore there was an edge in the dependengy graph that did not exist in
the parent dict. When the child item is completed, then its dependency ID is
looked up, but it doesn't exist, resulting in a crash.

By adding the _add_edge method, it should be easier to keep the edges of the
graph and the content of the dictionary in sync.