Do not hide console output when pressing : after a command
Normally, pressing ':' after executing a command would hide the console
(equivalent behavior to Escape or Enter). Now, : will focus the
commandline while keeping the previous output visible. This makes it
easier to compose a new command based on a todo shown in the previous
output (for example perform an 'ls' and then a 'dep' with IDs
afterwards).

Escape and Enter still have the same behavior, which can be used to
dismiss the last output.