Do not apply ordinal filtering when a tag appears more than once
Suppose we have the todo item:

    (C) Some item key:value1 key:value2

And then I do an ls:

    topydo ls -x key:value2

Before, this would yield an empty result, which is undesired. Therefore,
only apply ordinal tag filtering when a tag appears exactly once in the
todo item. This is a simple approach, and special operators >, >= etc
will not be applicable:

    topydo ls -x key:>value1

will no longer work. The result could be ambiguous: filter out when none of
the tags match, or only some of them?