Fix generic help and usage of invalid subcommands in aliases
Plain 'help' would crash the column UI, as would the usage of an invalid
command in an alias.

This case should be handled properly, as the CLI and Prompt mode do.

For this purpose, the generic help text was extracted, such that it can
be printed in column mode (and also in prompt mode, while I was at it).