Use Todo objs as args for Command in Transaction
This should fix InvalidTodoException when next todo id in Transaction
has been changed by previously finished Command.execute(). What we're
doing here is mimicking behavior of MultiCommand - we're getting Todo
objects and we supply them to the Command in place of the todo id.

This of course needed a change in TodoListBase.todo() to accept Todo
object as a potential argument for it.